fix(Main): bind global shortcut
